+++ client/src/features/common/TypeList.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { Col, Row } from 'react-bootstrap'
 
 import StyledDt from '../../styles/shared/DescriptionTerm'
 import StyledDd from '../../styles/shared/DescriptionDetail'
@@ -10,18 +11,20 @@ interface IProps {
 }
 
 const TypeList: React.FC<IProps> = ({ types }) => (
-  <React.Fragment>
-    <StyledDt data-testid="types-label">Categorized As</StyledDt>
-    <StyledDd data-testid="entity-type-list">
-      {types.map((type, index) => (
-        // eslint-disable-next-line react/no-array-index-key
-        <React.Fragment key={`${type}_${index}`}>
-          <RecordLink url={type} linkCategory="Results Snippet" />
-          {index === types.length - 1 ? '' : ', '}
-        </React.Fragment>
-      ))}
-    </StyledDd>
-  </React.Fragment>
+  <Row>
+    <Col>
+      <StyledDt data-testid="types-label">Categorized As</StyledDt>
+      <StyledDd data-testid="entity-type-list">
+        {types.map((type, index) => (
+          // eslint-disable-next-line react/no-array-index-key
+          <React.Fragment key={`${type}_${index}`}>
+            <RecordLink url={type} linkCategory="Results Snippet" />
+            {index === types.length - 1 ? '' : ', '}
+          </React.Fragment>
+        ))}
+      </StyledDd>
+    </Col>
+  </Row>
 )
 
 export default TypeList