+++ client/src/lib/util/params.ts
@@ -7,32 +7,27 @@ export const getParamPrefix = (tab: string): string =>
  * Function to format the requests for search estimates on each tab.
  * It returns an object with keys pertaining to the results tabs and their search criteria with any facets.
  * @param {Record<string, any>} criteria the search criteria from the q param in the url
- * @param {URLSearchParams} urlParams the search params from the current url
  * @returns {Record<string, string>}
  */
 export const getFacetParamsForSimpleSearchEstimatesRequest = (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  criteria: Record<string, any> | null,
+  criteria: Record<string, any>,
   urlParams: URLSearchParams,
-): Record<string, string> | null => {
-  if (criteria) {
-    const searchEstimatesParams: Record<string, string> = {}
-    Object.entries(searchScope).map(([key, value]) => {
-      const facetsParam = `${value.slice(0, 1)}f`
-      if (urlParams.has(facetsParam)) {
-        searchEstimatesParams[key] = `{"AND":[${JSON.stringify(
-          criteria,
-        )},${urlParams.get(facetsParam)}]}`
-      } else {
-        searchEstimatesParams[key] = JSON.stringify(criteria)
-      }
-      return null
-    })
+): Record<string, string> => {
+  const searchEstimatesParams: Record<string, string> = {}
+  Object.entries(searchScope).map(([key, value]) => {
+    const facetsParam = `${value.slice(0, 1)}f`
+    if (urlParams.has(facetsParam)) {
+      searchEstimatesParams[key] = `{"AND":[${JSON.stringify(
+        criteria,
+      )},${urlParams.get(facetsParam)}]}`
+    } else {
+      searchEstimatesParams[key] = JSON.stringify(criteria)
+    }
+    return null
+  })
 
-    return searchEstimatesParams
-  }
-
-  return null
+  return searchEstimatesParams
 }
 
 /**
@@ -44,22 +39,18 @@ export const getFacetParamsForSimpleSearchEstimatesRequest = (
  */
 export const getFacetParamsForAdvancedSearchEstimatesRequest = (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  criteria: Record<string, any> | null,
+  criteria: Record<string, any>,
   urlParams: URLSearchParams,
   queryTab: string,
-): string | null => {
-  if (criteria) {
-    let searchEstimatesParams = ''
-    const facetsParam = `${getParamPrefix(queryTab)}f`
-    if (urlParams.has(facetsParam)) {
-      searchEstimatesParams = `{"AND":[${JSON.stringify(
-        criteria,
-      )},${urlParams.get(facetsParam)}]}`
-    } else {
-      searchEstimatesParams = JSON.stringify(criteria)
-    }
-    return searchEstimatesParams
+): string => {
+  let searchEstimatesParams = ''
+  const facetsParam = `${getParamPrefix(queryTab)}f`
+  if (urlParams.has(facetsParam)) {
+    searchEstimatesParams = `{"AND":[${JSON.stringify(
+      criteria,
+    )},${urlParams.get(facetsParam)}]}`
+  } else {
+    searchEstimatesParams = JSON.stringify(criteria)
   }
-
-  return null
+  return searchEstimatesParams
 }