+++ client/src/test/unit/features/common/HowDoISeeIt.spec.tsx
@@ -1,13 +1,13 @@
 import { render } from '@testing-library/react'
 import React from 'react'
 
-import HowDoISeeIt from '../../../../features/works/HowDoISeeIt'
+import HowDoISeeIt from '../../../../features/common/HowDoISeeIt'
 import { reusableMinimalEntity as mockEntity } from '../../../data/reusableMinimalEntity'
 
 describe('HowDoISeeIt', () => {
   it('renders null', async () => {
     const { container } = render(
-      <HowDoISeeIt entity={mockEntity('Minimal Entity')} />,
+      <HowDoISeeIt data={mockEntity('Minimal Entity')} />,
     )
     expect(container.innerHTML).toBe('')
   })