+++ client/src/test/unit/lib/parse/search/estimatesParser.spec.ts
@@ -6,6 +6,7 @@ import {
 } from '../../../../../lib/parse/search/estimatesParser'
 import {
   advancedSearchEstimates,
+  estimatesErrorResults,
   estimatesResults,
   estimatesResultsWithRedirect,
   estimatesState,
@@ -20,7 +21,7 @@ describe('estimatesParser', () => {
       const mockEstimates = objectEstimates.objects
       const mockTransformedEstimates = advancedSearchEstimates
       expect(
-        transformAdvancedSearchEstimates(true, mockEstimates, 'objects'),
+        transformAdvancedSearchEstimates({ objects: mockEstimates }, 'objects'),
       ).toStrictEqual(mockTransformedEstimates)
     })
   })
@@ -29,7 +30,7 @@ describe('estimatesParser', () => {
     it('returns the translated estimates for all tabs', () => {
       const mockSimpleSearchEstimatesResults = estimatesResults
       expect(
-        transformSimpleSearchEstimates(true, mockSimpleSearchEstimatesResults),
+        transformSimpleSearchEstimates(mockSimpleSearchEstimatesResults),
       ).toStrictEqual({
         objects: 803,
         works: 100,
@@ -37,6 +38,20 @@ describe('estimatesParser', () => {
         places: 14,
         concepts: 730,
         events: 1,
+      })
+    })
+
+    it('returns the estimates when there are errors in the results', () => {
+      const mockSimpleSearchEstimatesResults = estimatesErrorResults
+      expect(
+        transformSimpleSearchEstimates(mockSimpleSearchEstimatesResults),
+      ).toStrictEqual({
+        objects: '-',
+        works: '-',
+        people: '-',
+        places: '-',
+        concepts: '-',
+        events: '-',
       })
     })
   })