+++ client/CHANGELOG.md
@@ -2,7 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.35.2 - 2024-10-2
+## v1.36.0 - Unreleased
+
+### Added
+
+- Added keyboard focus when navigating hierarchy ([#293](https://github.com/project-lux/lux-frontend/issues/293)).
+- Added access statement to works pages ([#364](https://github.com/project-lux/lux-frontend/issues/364)).
+
+### Changed
+
+### Removed
+
+### Fixed
+
+- Fixed layout of snippets if data fails to render ([#375](https://github.com/project-lux/lux-frontend/issues/375)).
+- Fixed professional activities rendering on people and group pages ([#321](https://github.com/project-lux/lux-frontend/issues/321)).
+
+## v1.35.2 - 2024-10-4
 
 ### Fixed